.agents/skills/financial_news/pyproject.toml
[project]
name = "financial-news"
version = "0.1.0"
description = "AI-powered financial news summarization dashboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "feedparser>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"
pythonpath = ["src"]
